Giving cl_test some love
This is a first step towards making a better test suite, one that detects what tests are compiled into it automatically instead of having to define them manually in an array, better macros, all tests wrapped in classes like in JL, etc.
The idea is to make it capable of hosting many tests, and to make the work of porting tests easier.